refactor: drop redundant casts and align internal construction
Tidy four internal spots the audit surfaced, none of which changes any output.

Centralize the half-open (0, 1] guard shared by the parabolic SAR's acceleration
factor / cap and MAMA's alpha limits into a validate_unit_fraction helper in the
_expr hub, replacing four byte-identical inline checks (same message, so the
raising tests are unchanged). Build the SuperTrend kernel's output struct
columnar via to_struct(), matching its two sibling kernels, instead of a
list-of-dicts. Drop three now-redundant casts: the Boolean rolling_max in
_rolling_has_nan (1.42 accepts a Boolean directly), the int-range before pl.corr
in stability, and the one before .sqrt() in probabilistic_sharpe_ratio (its
load-bearing Int64 cast stays). Note beside the two three-way window-ordering
guards that the hub helper is pairwise by design.

+++ src/pomata/metrics/performance.py
@@ -279,7 +279,7 @@ def stability(
     returns = float64_expr(returns)
     defined = returns.drop_nulls()
     cumulative_log = defined.log1p().cum_sum()
-    index = pl.int_range(0, defined.len(), dtype=pl.Int64).cast(pl.Float64)
+    index = pl.int_range(0, defined.len(), dtype=pl.Int64)
     # Clamp to the documented [0, 1] upper bound: a perfect linear fit gives a correlation that can round to 1 + eps,
     # whose square would otherwise escape the bound by a floating-point ulp.
     r_squared = (pl.corr(index, cumulative_log) ** 2).clip(upper_bound=1.0)
